test logic update
