+++ test/parse_points.hpp
@@ -12,6 +12,7 @@
 #include <memory>
 #include <stdexcept>
 #include <type_traits>
+#include <variant>
 
 #include "benchUtils.h"
 
@@ -287,9 +288,12 @@ inline auto load_point(const char *input_name, Conv converter, size_t max_num = 
   throw std::invalid_argument("Unsupported input spec");
 }
 
-template<typename L, typename pid_t = uint32_t>
-inline std::pair<std::vector<std::vector<L>>, std::unordered_map<L, std::vector<pid_t>>>
-load_label(const char *file_path, size_t max_size = 0) {
+// TODO: load labels from binary files
+template<typename L = uint32_t, typename pid_t = uint32_t>
+inline std::pair<std::vector<std::vector<L>>,
+                 std::variant<std::unordered_map<L, std::vector<pid_t>>,
+                              std::vector<std::pair<L, std::vector<pid_t>>>>>
+load_label(const char *file_path, size_t max_size = 0, bool ret_pair = true) {
   std::ifstream file(file_path);
   if (!file) {
     std::cerr << "Error: Unable to open file " << file_path << std::endl;
@@ -326,17 +330,7 @@ load_label(const char *file_path, size_t max_size = 0) {
     total_labels += node_labels.size();
     ++i;
   }
-
-  // std::sort(F.begin(), F.end(), [&](std::vector<L> a, std::vector<L> b) {
-  //   return a.size() > b.size();
-  // });
-  // size_t i = 0;
-  // for (const auto& labels : F) {
-  //   for (const L& label : labels) {
-  //     P[label].push_back(static_cast<pid_t>(i));
-  //   }
-  //   ++i;
-  // }
+  file.close();
 
   std::cout << std::fixed << std::setprecision(2)
             << "Filters per Point: " << (float)total_labels / (float)num_points << std::endl;
@@ -348,14 +342,18 @@ load_label(const char *file_path, size_t max_size = 0) {
       [&](const std::pair<L, std::vector<pid_t>> &a, const std::pair<L, std::vector<pid_t>> &b) {
         return a.second.size() > b.second.size();
       });
-  
+
+  if (ret_pair) {
+    std::reverse(pairs.begin(), pairs.end());
+    return std::make_pair(F, pairs);
+  }
+
   P.clear();
   // P = std::unordered_map<L, std::vector<pid_t>>(pairs.rbegin(), pairs.rend());
   for (const auto &it : pairs) {
     P.insert(it);
   }
 
-  file.close();
   return std::make_pair(F, P);
 }
